# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface RegisterInput {
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
  name?: string;
}

export interface LoginInput {
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface Entry {
  id: number;
  date: string;
  text: string;
  createdAt: string;
}

export interface EntryInput {
  /** @minLength 1 */
  date: string;
  text: string;
}

export type QuoteSourceType = typeof QuoteSourceType[keyof typeof QuoteSourceType];


export const QuoteSourceType = {
  journal: 'journal',
  saved_quote: 'saved_quote',
  copied_text: 'copied_text',
  unknown: 'unknown',
} as const;

export interface Quote {
  date: string;
  fragment: string;
  source_type: QuoteSourceType;
}

export type CandidateMode = typeof CandidateMode[keyof typeof CandidateMode];


export const CandidateMode = {
  thread: 'thread',
  memory: 'memory',
  distance: 'distance',
  value_signal: 'value_signal',
  wisdom: 'wisdom',
} as const;

export interface Candidate {
  mode: CandidateMode;
  candidate_title?: string;
  function: string;
  description: string;
  evidence: Quote[];
  why_it_matters: string;
}

export interface ExtractInput {
  entries: string;
}

export interface ExtractResult {
  candidates: Candidate[];
}

export interface CandidateScore {
  mode: string;
  candidate_title?: string;
  safety: number;
  worth_returning_to: number;
  recognition: number;
  specificity: number;
  /** @nullable */
  persistence?: number | null;
  /** @nullable */
  same_function_different_language?: number | null;
  /** @nullable */
  vividness?: number | null;
  /** @nullable */
  revealing?: number | null;
  /** @nullable */
  contrast?: number | null;
  /** @nullable */
  evidence_across_time?: number | null;
  /** @nullable */
  meaningfulness_of_selection?: number | null;
  /** @nullable */
  attribution_confidence?: number | null;
  /** @nullable */
  clarity?: number | null;
  /** @nullable */
  earnedness?: number | null;
  surfaceable: boolean;
  why: string;
}

export interface ScoreInput {
  candidates: Candidate[];
}

export type ScoreResultMode = typeof ScoreResultMode[keyof typeof ScoreResultMode];


export const ScoreResultMode = {
  thread: 'thread',
  memory: 'memory',
  distance: 'distance',
  value_signal: 'value_signal',
  wisdom: 'wisdom',
  nothing: 'nothing',
} as const;

export interface ScoreResult {
  scores: CandidateScore[];
  mode: ScoreResultMode;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  observation?: string | null;
  quotes: Quote[];
  why: string;
  /** @nullable */
  message?: string | null;
}